refactor: AnnotationHandlerAdapter 의 handle() 메서드 수정
